Add NoopFormatter
This is useful when using a custom logger: with the LogTarget. For
example, a `Logfmt::Logger` logger will take a Hash and correctly format
it to the Logfmt format. When using that logger in a Rails app, this
allows this plugin's config to look like this:

```ruby
Puma::Plugin::Telemetry.configure do |config|
  config.add_target(:log, logger: Rails.logger, transform: :l2met, formatter: :noop)
end
```

see: https://github.com/cyberdelia/logfmt-ruby?tab=readme-ov-file#writing-log-lines