Bring back Gemfile.lock and use latest Ruby 2.6
Even though Ruby 2.6 (and 2.7) are EoL'd, they are still supported by
this Gem. So let's use that oldest one locally too. Also, it appears it
is now recommended to include the Gemfile.lock in Git for Gems. The
guidance from Bundler use to be _NOT_ to include the .lock file. But,
time moves forward and things change. I'll adapt and learn too! 😅

https://bundler.io/guides/faq.html#using-gemfiles-inside-gems