Cleanup gem setup (#20)
* Ignore Gemfile.lock from Git

We don't really want to lock folks working on this gem into exact
versions of the various dependencies. Especially for  certain
dependencies with native extensions, which might not install on
different platforms. This also organizes the .gitignore.

* Exclude dev-specific files from packed Gem

We can keep the size of the packaged Gem smaller by no including specs,
binstubs, git/github config, and other development-time-only files.

* Upgrade Rake (b/c why not?)

* Fix some spelling, grammar, and formatting

* Test newer Rubies and Ubuntus too!

* fixup ubuntu

* Run tests on a new PR

* Bring back Gemfile.lock and use latest Ruby 2.6

Even though Ruby 2.6 (and 2.7) are EoL'd, they are still supported by
this Gem. So let's use that oldest one locally too. Also, it appears it
is now recommended to include the Gemfile.lock in Git for Gems. The
guidance from Bundler use to be _NOT_ to include the .lock file. But,
time moves forward and things change. I'll adapt and learn too! 😅

https://bundler.io/guides/faq.html#using-gemfiles-inside-gems

* Appease Rubocop 👮